requirement.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
httpx
google-generativeai
google-genai
sentence-transformers
faiss-cpu
torch
torchvision
torchaudio
scikit-learn
numpy
pandas
tqdm
requests
google-adk
google-generativeai
google-genai
python-dotenv
ipython